Refactors * Common functionality to base classes * GT Pre/Postload broken into more modular pieces
