+++ src/main/java/gregtech/api/metatileentity/BaseMetaTileEntity.java
@@ -85,12 +85,17 @@ public class BaseMetaTileEntity extends BaseTileEntity implements IGregTechTileE
     protected int mAverageEUInputIndex = 0, mAverageEUOutputIndex = 0;
     protected boolean mReleaseEnergy = false;
     protected long[] mAverageEUInput = new long[]{0, 0, 0, 0, 0}, mAverageEUOutput = new long[]{0, 0, 0, 0, 0};
-    private boolean[] mActiveEUInputs = new boolean[]{false, false, false, false, false, false}, mActiveEUOutputs = new boolean[]{false, false, false, false, false, false};
+    private final boolean[] mActiveEUInputs = new boolean[]{false, false, false, false, false, false};
+    private final boolean[] mActiveEUOutputs = new boolean[]{false, false, false, false, false, false};
     private byte[] mSidedRedstone = new byte[]{15, 15, 15, 15, 15, 15};
-    private int[] mCoverSides = new int[]{0, 0, 0, 0, 0, 0}, mTimeStatistics = new int[GregTech_API.TICKS_FOR_LAG_AVERAGING];
+    private int[] mCoverSides = new int[]{0, 0, 0, 0, 0, 0};
+    private final int[] mTimeStatistics = new int[GregTech_API.TICKS_FOR_LAG_AVERAGING];
     private ISerializableObject[] mCoverData = new ISerializableObject[6];
-    private boolean mHasEnoughEnergy = true, mRunningThroughTick = false, mInputDisabled = false, mOutputDisabled = false, mMuffler = false, mLockUpgrade = false, mActive = false, mRedstone = false, mWorkUpdate = false, mSteamConverter = false, mInventoryChanged = false, mWorks = true, mNeedsUpdate = true, mNeedsBlockUpdate = true, mSendClientData = false, oRedstone = false;
-    private byte mColor = 0, oColor = 0, oStrongRedstone = 0, mStrongRedstone = 0, oRedstoneData = 63, oTextureData = 0, oUpdateData = 0, oTexturePage=0, oLightValueClient = -1, oLightValue = -1, mLightValue = 0, mOtherUpgrades = 0, mFacing = 0, oFacing = 0, mWorkData = 0;
+    private boolean mHasEnoughEnergy = true, mRunningThroughTick = false, mInputDisabled = false, mOutputDisabled = false, mMuffler = false, mLockUpgrade = false; 
+    private boolean mActive = false, mRedstone = false, mWorkUpdate = false, mSteamConverter = false, mInventoryChanged = false, mWorks = true, mNeedsUpdate = true;
+    private boolean mNeedsBlockUpdate = true, mSendClientData = false, oRedstone = false;
+    private byte mColor = 0, oColor = 0, oStrongRedstone = 0, mStrongRedstone = 0, oRedstoneData = 63, oTextureData = 0, oUpdateData = 0, oTexturePage=0;
+    private byte oLightValueClient = -1, oLightValue = -1, mLightValue = 0, mOtherUpgrades = 0, mFacing = 0, oFacing = 0, mWorkData = 0;
     private int mDisplayErrorCode = 0, oX = 0, oY = 0, oZ = 0, mTimeStatisticsIndex = 0, mLagWarningCount = 0;
     private short mID = 0;
     public long mTickTimer = 0;
@@ -132,7 +137,7 @@ public class BaseMetaTileEntity extends BaseTileEntity implements IGregTechTileE
         try {
             super.writeToNBT(aNBT);
         } catch (Throwable e) {
-            GT_Log.err.println("Encountered CRITICAL ERROR while saving MetaTileEntity, the Chunk whould've been corrupted by now, but I prevented that. Please report immediately to GregTech Intergalactical!!!");
+            GT_Log.err.println("Encountered CRITICAL ERROR while saving MetaTileEntity.");
             e.printStackTrace(GT_Log.err);
         }
         try {
@@ -164,7 +169,7 @@ public class BaseMetaTileEntity extends BaseTileEntity implements IGregTechTileE
             aNBT.setTag("GT.CraftingComponents", mRecipeStuff);
             aNBT.setInteger("nbtVersion", GT_Mod.TOTAL_VERSION);
         } catch (Throwable e) {
-            GT_Log.err.println("Encountered CRITICAL ERROR while saving MetaTileEntity, the Chunk whould've been corrupted by now, but I prevented that. Please report immediately to GregTech Intergalactical!!!");
+            GT_Log.err.println("Encountered CRITICAL ERROR while saving MetaTileEntity.");
             e.printStackTrace(GT_Log.err);
         }
         try {
@@ -184,12 +189,12 @@ public class BaseMetaTileEntity extends BaseTileEntity implements IGregTechTileE
                 try {
                     mMetaTileEntity.saveNBTData(aNBT);
                 } catch (Throwable e) {
-                    GT_Log.err.println("Encountered CRITICAL ERROR while saving MetaTileEntity, the Chunk whould've been corrupted by now, but I prevented that. Please report immediately to GregTech Intergalactical!!!");
+                    GT_FML_LOGGER.error("Encountered CRITICAL ERROR while saving MetaTileEntity.");
                     e.printStackTrace(GT_Log.err);
                 }
             }
         } catch (Throwable e) {
-            GT_Log.err.println("Encountered CRITICAL ERROR while saving MetaTileEntity, the Chunk whould've been corrupted by now, but I prevented that. Please report immediately to GregTech Intergalactical!!!");
+            GT_Log.err.println("Encountered CRITICAL ERROR while saving MetaTileEntity.");
             e.printStackTrace(GT_Log.err);
         }
     }
@@ -341,38 +346,259 @@ public class BaseMetaTileEntity extends BaseTileEntity implements IGregTechTileE
 
         mRunningThroughTick = true;
         long tTime = System.nanoTime();
-        int tCode = 0;
         boolean aSideServer = isServerSide();
         boolean aSideClient = isClientSide();
 
-        try { for (tCode = 0; hasValidMetaTileEntity() && tCode >= 0; ) {
-            switch (tCode) {
-                case 0:
-                    tCode++;
-                    if (mTickTimer++ == 0) {
+        try {
+            if (hasValidMetaTileEntity()) {
+                if (mTickTimer++ == 0) {
+                    oX = xCoord;
+                    oY = yCoord;
+                    oZ = zCoord;
+                    if (aSideServer) {
+                        checkDropCover();
+                    }
+                    worldObj.markTileEntityChunkModified(xCoord, yCoord, zCoord, this);
+                    mMetaTileEntity.onFirstTick(this);
+                    if (!hasValidMetaTileEntity()) {
+                        mRunningThroughTick = false;
+                        return;
+                    }
+                }
+                if (aSideClient) {
+                    if (mColor != oColor) {
+                        mMetaTileEntity.onColorChangeClient(oColor = mColor);
+                        issueTextureUpdate();
+                    }
+
+                    if (mLightValue != oLightValueClient) {
+                        worldObj.setLightValue(EnumSkyBlock.Block, xCoord, yCoord, zCoord, mLightValue);
+                        worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord, zCoord);
+                        worldObj.updateLightByType(EnumSkyBlock.Block, xCoord + 1, yCoord, zCoord);
+                        worldObj.updateLightByType(EnumSkyBlock.Block, xCoord - 1, yCoord, zCoord);
+                        worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord + 1, zCoord);
+                        worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord - 1, zCoord);
+                        worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord, zCoord + 1);
+                        worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord, zCoord - 1);
+                        oLightValueClient = mLightValue;
+                        issueTextureUpdate();
+                    }
+
+                    if (mNeedsUpdate) {
+                        worldObj.markBlockForUpdate(xCoord, yCoord, zCoord);
+                        //worldObj.func_147479_m(xCoord, yCoord, zCoord);
+                        mNeedsUpdate = false;
+                    }
+                }
+                if (aSideServer && mTickTimer > 10) {
+                    for (byte i = 0; i < 6; i++)
+                        if (getCoverIDAtSide(i) != 0) {
+                            GT_CoverBehaviorBase<?> tCover = getCoverBehaviorAtSideNew(i);
+                            int tCoverTickRate = tCover.getTickRate(i, getCoverIDAtSide(i), mCoverData[i], this);
+                            if (tCoverTickRate > 0 && mTickTimer % tCoverTickRate == 0) {
+                                byte tRedstone = tCover.isRedstoneSensitive(i, getCoverIDAtSide(i), mCoverData[i], this, mTickTimer) ? getInputRedstoneSignal(i) : 0;
+                                mCoverData[i] = tCover.doCoverThings(i, tRedstone, getCoverIDAtSide(i), mCoverData[i], this, mTickTimer);
+                                if (!hasValidMetaTileEntity()) {
+                                    mRunningThroughTick = false;
+                                    return;
+                                }
+                            }
+                        }
+
+                }
+                if (aSideServer) {
+                    if (++mAverageEUInputIndex >= mAverageEUInput.length) mAverageEUInputIndex = 0;
+                    if (++mAverageEUOutputIndex >= mAverageEUOutput.length) mAverageEUOutputIndex = 0;
+
+                    mAverageEUInput[mAverageEUInputIndex] = 0;
+                    mAverageEUOutput[mAverageEUOutputIndex] = 0;
+                }
+                
+                mMetaTileEntity.onPreTick(this, mTickTimer);
+                
+                if (!hasValidMetaTileEntity()) {
+                    mRunningThroughTick = false;
+                    return;
+                }
+                if (aSideServer) {
+                    if (mRedstone != oRedstone || mTickTimer == 10) {
+                        for (byte i = 0; i < 6; i++)
+                            mCoverBehaviors[i] = GregTech_API.getCoverBehaviorNew(mCoverSides[i]);
+                        oRedstone = mRedstone;
+                        issueBlockUpdate();
+                    }
+                    if (mTickTimer == 10) joinEnet();
+
+                    if (xCoord != oX || yCoord != oY || zCoord != oZ) {
                         oX = xCoord;
                         oY = yCoord;
                         oZ = zCoord;
-                        if (aSideServer) for (byte i = 0; i < 6; i++)
-                            if (getCoverIDAtSide(i) != 0)
-                                if (!mMetaTileEntity.allowCoverOnSide(i, new GT_ItemStack(getCoverIDAtSide(i))))
-                                    dropCover(i, i, true);
-                        worldObj.markTileEntityChunkModified(xCoord, yCoord, zCoord, this);
-                        mMetaTileEntity.onFirstTick(this);
-                        if (!hasValidMetaTileEntity()) {
-                            mRunningThroughTick = false;
-                            return;
-                        }
+                        issueClientUpdate();
+                        clearTileEntityBuffer();
                     }
-                case 1:
-                    tCode++;
-                    if (aSideClient) {
-                        if (mColor != oColor) {
-                            mMetaTileEntity.onColorChangeClient(oColor = mColor);
-                            issueTextureUpdate();
+
+                    if (mFacing != oFacing) {
+                        oFacing = mFacing;
+                        checkDropCover();
+                        issueBlockUpdate();
+                    }
+
+                    if (mTickTimer > 20 && mMetaTileEntity.isElectric()) {
+                        mAcceptedAmperes = 0;
+
+                        if (getOutputVoltage() != oOutput) {
+                            oOutput = getOutputVoltage();
                         }
 
-                        if (mLightValue != oLightValueClient) {
+                        if (mMetaTileEntity.isEnetOutput() || mMetaTileEntity.isEnetInput()) {
+                            for (byte i = 0; i < 6; i++) {
+                                boolean temp = isEnergyInputSide(i);
+                                if (temp != mActiveEUInputs[i]) {
+                                    mActiveEUInputs[i] = temp;
+                                }
+                                temp = isEnergyOutputSide(i);
+                                if (temp != mActiveEUOutputs[i]) {
+                                    mActiveEUOutputs[i] = temp;
+                                }
+                            }
+                        }
+
+
+                        if (mMetaTileEntity.isEnetOutput() && oOutput > 0) {
+                            long tOutputVoltage = Math.max(oOutput, oOutput + (1L << Math.max(0, GT_Utility.getTier(oOutput) - 1))),
+                                tUsableAmperage = Math.min(getOutputAmperage(), (getStoredEU() - mMetaTileEntity.getMinimumStoredEU()) / tOutputVoltage);
+                            if (tUsableAmperage > 0) {
+                                long tEU = tOutputVoltage * Util.emitEnergyToNetwork(oOutput, tUsableAmperage, this);
+                                mAverageEUOutput[mAverageEUOutputIndex] += tEU;
+                                decreaseStoredEU(tEU, true);
+                            }
+                        }
+                        if (getEUCapacity() > 0) {
+                            if (GregTech_API.sMachineFireExplosions && getRandomNumber(1000) == 0) {
+                                Block tBlock = getBlockAtSide((byte) getRandomNumber(6));
+                                if (tBlock instanceof BlockFire) doEnergyExplosion();
+                            }
+
+                            if (!hasValidMetaTileEntity()) {
+                                mRunningThroughTick = false;
+                                return;
+                            }
+
+                            if (getRandomNumber(1000) == 0) {
+                                int precipitationHeightAtSide2 = worldObj.getPrecipitationHeight(xCoord, zCoord - 1);
+                                int precipitationHeightAtSide3 = worldObj.getPrecipitationHeight(xCoord, zCoord + 1);
+                                int precipitationHeightAtSide4 = worldObj.getPrecipitationHeight(xCoord - 1, zCoord);
+                                int precipitationHeightAtSide5 = worldObj.getPrecipitationHeight(xCoord + 1, zCoord);
+
+                                if ((getCoverIDAtSide((byte) 1) == 0 && worldObj.getPrecipitationHeight(xCoord, zCoord) - 2 < yCoord)
+                                    || (getCoverIDAtSide((byte) 2) == 0 && precipitationHeightAtSide2 - 1 < yCoord && precipitationHeightAtSide2 > -1)
+                                    || (getCoverIDAtSide((byte) 3) == 0 && precipitationHeightAtSide3 - 1 < yCoord && precipitationHeightAtSide3 > -1)
+                                    || (getCoverIDAtSide((byte) 4) == 0 && precipitationHeightAtSide4 - 1 < yCoord && precipitationHeightAtSide4 > -1)
+                                    || (getCoverIDAtSide((byte) 5) == 0 && precipitationHeightAtSide5 - 1 < yCoord && precipitationHeightAtSide5 > -1)) {
+                                    if (GregTech_API.sMachineRainExplosions && worldObj.isRaining() && getBiome().rainfall > 0) {
+                                        if (getRandomNumber(10) == 0) {
+                                            try {
+                                                GT_Mod.instance.achievements.issueAchievement(this.getWorldObj().getPlayerEntityByName(mOwnerName), "badweather");
+                                            } catch (Exception ignored) {}
+                                            GT_Log.exp.println("Machine at: " + this.getXCoord() + " | " + this.getYCoord() + " | " + this.getZCoord() + " DIMID: " + this.worldObj.provider.dimensionId + " explosion due to rain!");
+                                            doEnergyExplosion();
+                                        } else {
+                                            GT_Log.exp.println("Machine at: " + this.getXCoord() + " | " + this.getYCoord() + " | " + this.getZCoord() + " DIMID: " + this.worldObj.provider.dimensionId + "  set to Fire due to rain!");
+                                            setOnFire();
+                                        }
+                                    }
+                                    if (!hasValidMetaTileEntity()) {
+                                        mRunningThroughTick = false;
+                                        return;
+                                    }
+                                    if (GregTech_API.sMachineThunderExplosions && worldObj.isThundering() && getBiome().rainfall > 0 && getRandomNumber(
+                                        3) == 0) {
+                                        try {
+                                            GT_Mod.instance.achievements.issueAchievement(this.getWorldObj().getPlayerEntityByName(mOwnerName), "badweather");
+                                        } catch (Exception ignored) {}
+                                        GT_Log.exp.println("Machine at: " + this.getXCoord() + " | " + this.getYCoord() + " | " + this.getZCoord() + " DIMID: " + this.worldObj.provider.dimensionId + " explosion due to Thunderstorm!");
+                                        doEnergyExplosion();
+                                    }
+                                }
+                            }
+                        }
+                    }
+
+                    if (!hasValidMetaTileEntity()) {
+                        mRunningThroughTick = false;
+                        return;
+                    }
+                }
+                if (aSideServer) {
+                    if (mMetaTileEntity.dechargerSlotCount() > 0 && getStoredEU() < getEUCapacity()) {
+                        for (int i = mMetaTileEntity.dechargerSlotStartIndex(), k = mMetaTileEntity.dechargerSlotCount() + i; i < k; i++) {
+                            if (mMetaTileEntity.mInventory[i] != null && getStoredEU() < getEUCapacity()) {
+                                dischargeItem(mMetaTileEntity.mInventory[i]);
+                                if (ic2.api.info.Info.itemEnergy.getEnergyValue(mMetaTileEntity.mInventory[i]) > 0) {
+                                    if ((getStoredEU() + ic2.api.info.Info.itemEnergy.getEnergyValue(mMetaTileEntity.mInventory[i])) < getEUCapacity()) {
+                                        increaseStoredEnergyUnits((long) ic2.api.info.Info.itemEnergy.getEnergyValue(mMetaTileEntity.mInventory[i]), false);
+                                        mMetaTileEntity.mInventory[i].stackSize--;
+                                        mInventoryChanged = true;
+                                    }
+                                }
+                                if (mMetaTileEntity.mInventory[i].stackSize <= 0) {
+                                    mMetaTileEntity.mInventory[i] = null;
+                                    mInventoryChanged = true;
+                                }
+                            }
+                        }
+                    }
+                }
+                if (aSideServer) {
+                    if (mMetaTileEntity.rechargerSlotCount() > 0 && getStoredEU() > 0) {
+                        for (int i = mMetaTileEntity.rechargerSlotStartIndex(), k = mMetaTileEntity.rechargerSlotCount() + i; i < k; i++) {
+                            if (getStoredEU() > 0 && mMetaTileEntity.mInventory[i] != null) {
+                                chargeItem(mMetaTileEntity.mInventory[i]);
+                                if (mMetaTileEntity.mInventory[i].stackSize <= 0) {
+                                    mMetaTileEntity.mInventory[i] = null;
+                                    mInventoryChanged = true;
+                                }
+                            }
+                        }
+                    }
+                }
+                updateStatus();
+                if (!hasValidMetaTileEntity()) {
+                    mRunningThroughTick = false;
+                    return;
+                }
+                mMetaTileEntity.onPostTick(this, mTickTimer);
+                if (!hasValidMetaTileEntity()) {
+                    mRunningThroughTick = false;
+                    return;
+                }
+                if (aSideServer) {
+                    if (mTickTimer > 20 && cableUpdateDelay == 0) {
+                        generatePowerNodes();
+                    }
+                    cableUpdateDelay--;
+                    if (mTickTimer % 10 == 0) {
+                        sendClientData();
+                    }
+
+                    if (mTickTimer > 10) {
+                        byte tData = (byte) ((mFacing & 7) | (mActive ? 8 : 0) | (mRedstone ? 16 : 0) | (mLockUpgrade ? 32 : 0) | (mWorks ? 64 : 0));
+                        if (tData != oTextureData)
+                            sendBlockEvent(ClientEvents.CHANGE_COMMON_DATA, oTextureData = tData);
+
+                        tData = mMetaTileEntity.getUpdateData();
+                        if (tData != oUpdateData)
+                            sendBlockEvent(ClientEvents.CHANGE_CUSTOM_DATA, oUpdateData = tData);
+                        if (mMetaTileEntity instanceof GT_MetaTileEntity_Hatch) {
+                            tData = ((GT_MetaTileEntity_Hatch) mMetaTileEntity).getTexturePage();
+                            if (tData != oTexturePage)
+                                sendBlockEvent(ClientEvents.CHANGE_CUSTOM_DATA, (byte) ((oTexturePage = tData) | 0x80));//set last bit as a flag for page
+                        }
+                        if (mColor != oColor) sendBlockEvent(ClientEvents.CHANGE_COLOR, oColor = mColor);
+                        tData = (byte) (((mSidedRedstone[0] > 0) ? 1 : 0) | ((mSidedRedstone[1] > 0) ? 2 : 0) | ((mSidedRedstone[2] > 0) ? 4 : 0) | ((mSidedRedstone[3] > 0) ? 8 : 0) | ((mSidedRedstone[4] > 0) ? 16 : 0) | ((mSidedRedstone[5] > 0) ? 32 : 0));
+                        if (tData != oRedstoneData)
+                            sendBlockEvent(ClientEvents.CHANGE_REDSTONE_OUTPUT, oRedstoneData = tData);
+                        if (mLightValue != oLightValue) {
                             worldObj.setLightValue(EnumSkyBlock.Block, xCoord, yCoord, zCoord, mLightValue);
                             worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord, zCoord);
                             worldObj.updateLightByType(EnumSkyBlock.Block, xCoord + 1, yCoord, zCoord);
@@ -381,291 +607,19 @@ public class BaseMetaTileEntity extends BaseTileEntity implements IGregTechTileE
                             worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord - 1, zCoord);
                             worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord, zCoord + 1);
                             worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord, zCoord - 1);
-                            oLightValueClient = mLightValue;
                             issueTextureUpdate();
-                        }
-
-                        if (mNeedsUpdate) {
-                            worldObj.markBlockForUpdate(xCoord, yCoord, zCoord);
-                            //worldObj.func_147479_m(xCoord, yCoord, zCoord);
-                            mNeedsUpdate = false;
+                            sendBlockEvent(ClientEvents.CHANGE_LIGHT, oLightValue = mLightValue);
                         }
                     }
-                case 2:
-                case 3:
-                case 4:
-                case 5:
-                case 6:
-                case 7:
-                    if (aSideServer && mTickTimer > 10) {
-                        for (byte i = (byte) (tCode - 2); i < 6; i++)
-                            if (getCoverIDAtSide(i) != 0) {
-                                tCode++;
-                                GT_CoverBehaviorBase<?> tCover = getCoverBehaviorAtSideNew(i);
-                                int tCoverTickRate = tCover.getTickRate(i, getCoverIDAtSide(i), mCoverData[i], this);
-                                if (tCoverTickRate > 0 && mTickTimer % tCoverTickRate == 0) {
-                                    byte tRedstone = tCover.isRedstoneSensitive(i, getCoverIDAtSide(i), mCoverData[i], this, mTickTimer) ? getInputRedstoneSignal(i) : 0;
-                                    mCoverData[i] = tCover.doCoverThings(i, tRedstone, getCoverIDAtSide(i), mCoverData[i], this, mTickTimer);
-                                    if (!hasValidMetaTileEntity()) {
-                                        mRunningThroughTick = false;
-                                        return;
-                                    }
-                                }
-                            }
 
+                    if (mNeedsBlockUpdate) {
+                        updateNeighbours(mStrongRedstone, oStrongRedstone);
+                        oStrongRedstone = mStrongRedstone;
+                        mNeedsBlockUpdate = false;
                     }
-                case 8:
-                    tCode = 9;
-                    if (aSideServer) {
-                        if (++mAverageEUInputIndex >= mAverageEUInput.length) mAverageEUInputIndex = 0;
-                        if (++mAverageEUOutputIndex >= mAverageEUOutput.length) mAverageEUOutputIndex = 0;
-
-                        mAverageEUInput[mAverageEUInputIndex] = 0;
-                        mAverageEUOutput[mAverageEUOutputIndex] = 0;
-                    }
-                case 9:
-                    tCode++;
-                    mMetaTileEntity.onPreTick(this, mTickTimer);
-                    if (!hasValidMetaTileEntity()) {
-                        mRunningThroughTick = false;
-                        return;
-                    }
-                case 10:
-                    tCode++;
-                    if (aSideServer) {
-                        if (mRedstone != oRedstone || mTickTimer == 10) {
-                            for (byte i = 0; i < 6; i++)
-                                mCoverBehaviors[i] = GregTech_API.getCoverBehaviorNew(mCoverSides[i]);
-                            oRedstone = mRedstone;
-                            issueBlockUpdate();
-                        }
-                        if(mTickTimer == 10) joinEnet();
-
-                        if (xCoord != oX || yCoord != oY || zCoord != oZ) {
-                            oX = xCoord;
-                            oY = yCoord;
-                            oZ = zCoord;
-                            issueClientUpdate();
-                            clearTileEntityBuffer();
-                        }
-
-                        if (mFacing != oFacing) {
-                            oFacing = mFacing;
-                            for (byte i = 0; i < 6; i++)
-                                if (getCoverIDAtSide(i) != 0)
-                                    if (!mMetaTileEntity.allowCoverOnSide(i, new GT_ItemStack(getCoverIDAtSide(i))))
-                                        dropCover(i, i, true);
-                            issueBlockUpdate();
-                        }
-
-                        if (mTickTimer > 20 && mMetaTileEntity.isElectric()) {
-                            mAcceptedAmperes = 0;
-
-                            if (getOutputVoltage() != oOutput) {
-                                oOutput = getOutputVoltage();
-                            }
-
-                            if (mMetaTileEntity.isEnetOutput() || mMetaTileEntity.isEnetInput()) {
-                                for (byte i = 0; i < 6; i++) {
-                                    boolean
-                                            temp = isEnergyInputSide(i);
-                                    if (temp != mActiveEUInputs[i]) {
-                                        mActiveEUInputs[i] = temp;
-                                    }
-                                    temp = isEnergyOutputSide(i);
-                                    if (temp != mActiveEUOutputs[i]) {
-                                        mActiveEUOutputs[i] = temp;
-                                    }
-                                }
-                            }
-
-
-                            if (mMetaTileEntity.isEnetOutput() && oOutput > 0) {
-                                long tOutputVoltage = Math.max(oOutput, oOutput + (1L << Math.max(0, GT_Utility.getTier(oOutput) - 1))),
-                                        tUsableAmperage = Math.min(getOutputAmperage(), (getStoredEU() - mMetaTileEntity.getMinimumStoredEU()) / tOutputVoltage);
-                                if (tUsableAmperage > 0) {
-                                    long tEU = tOutputVoltage * IEnergyConnected.Util.emitEnergyToNetwork(oOutput, tUsableAmperage, this);
-                                    mAverageEUOutput[mAverageEUOutputIndex] += tEU;
-                                    decreaseStoredEU(tEU, true);
-                                }
-                            }
-                            if (getEUCapacity() > 0) {
-                                if (GregTech_API.sMachineFireExplosions && getRandomNumber(1000) == 0) {
-                                    Block tBlock = getBlockAtSide((byte) getRandomNumber(6));
-                                    if (tBlock instanceof BlockFire) doEnergyExplosion();
-                                }
-
-                                if (!hasValidMetaTileEntity()) {
-                                    mRunningThroughTick = false;
-                                    return;
-                                }
-
-                                if (getRandomNumber(1000) == 0) {
-				    int precipitationHeightAtSide2 = worldObj.getPrecipitationHeight(xCoord, zCoord - 1);
-                                    int precipitationHeightAtSide3 = worldObj.getPrecipitationHeight(xCoord, zCoord + 1);
-                                    int precipitationHeightAtSide4 = worldObj.getPrecipitationHeight(xCoord - 1, zCoord);
-                                    int precipitationHeightAtSide5 = worldObj.getPrecipitationHeight(xCoord + 1, zCoord);
-
-                                    if ((getCoverIDAtSide((byte) 1) == 0 && worldObj.getPrecipitationHeight(xCoord, zCoord) - 2 < yCoord)
-                                            || (getCoverIDAtSide((byte) 2) == 0 && precipitationHeightAtSide2 - 1 < yCoord && precipitationHeightAtSide2 > -1)
-                                            || (getCoverIDAtSide((byte) 3) == 0 && precipitationHeightAtSide3 - 1 < yCoord && precipitationHeightAtSide3 > -1)
-                                            || (getCoverIDAtSide((byte) 4) == 0 && precipitationHeightAtSide4 - 1 < yCoord && precipitationHeightAtSide4 > -1)
-                                            || (getCoverIDAtSide((byte) 5) == 0 && precipitationHeightAtSide5 - 1 < yCoord && precipitationHeightAtSide5 > -1)) {
-                                        if (GregTech_API.sMachineRainExplosions && worldObj.isRaining() && getBiome().rainfall > 0) {
-                                            if (getRandomNumber(10) == 0) {
-                                                try{
-                                                    GT_Mod.instance.achievements.issueAchievement(this.getWorldObj().getPlayerEntityByName(mOwnerName), "badweather");
-                                                }catch(Exception e){
-
-                                                }
-                                                GT_Log.exp.println("Machine at: "+ this.getXCoord()+" | "+ this.getYCoord()+" | "+ this.getZCoord()+" DIMID: " +this.worldObj.provider.dimensionId +" explosion due to rain!");
-                                                doEnergyExplosion();
-                                            } else {
-                                                GT_Log.exp.println("Machine at: "+ this.getXCoord()+" | "+ this.getYCoord()+" | "+ this.getZCoord()+" DIMID: " +this.worldObj.provider.dimensionId +"  set to Fire due to rain!");
-                                                setOnFire();
-                                            }
-                                        }
-                                        if (!hasValidMetaTileEntity()) {
-                                            mRunningThroughTick = false;
-                                            return;
-                                        }
-                                        if (GregTech_API.sMachineThunderExplosions && worldObj.isThundering() && getBiome().rainfall > 0 && getRandomNumber(3) == 0) {
-                                        	try{
-                                        	    GT_Mod.instance.achievements.issueAchievement(this.getWorldObj().getPlayerEntityByName(mOwnerName), "badweather");
-                                        	}catch(Exception e){
-
-                                            }
-                                            GT_Log.exp.println("Machine at: "+ this.getXCoord()+" | "+ this.getYCoord()+" | "+ this.getZCoord()+" DIMID: " +this.worldObj.provider.dimensionId + " explosion due to Thunderstorm!");
-                                            doEnergyExplosion();
-                                        }
-                                    }
-                                }
-                            }
-                        }
-
-                        if (!hasValidMetaTileEntity()) {
-                            mRunningThroughTick = false;
-                            return;
-                        }
-                    }
-                case 11:
-                    tCode++;
-                    if (aSideServer) {
-                        if (mMetaTileEntity.dechargerSlotCount() > 0 && getStoredEU() < getEUCapacity()) {
-                            for (int i = mMetaTileEntity.dechargerSlotStartIndex(), k = mMetaTileEntity.dechargerSlotCount() + i; i < k; i++) {
-                                if (mMetaTileEntity.mInventory[i] != null && getStoredEU() < getEUCapacity()) {
-                                    dischargeItem(mMetaTileEntity.mInventory[i]);
-				                    if(ic2.api.info.Info.itemEnergy.getEnergyValue(mMetaTileEntity.mInventory[i])>0){
-                                       if((getStoredEU() + ic2.api.info.Info.itemEnergy.getEnergyValue(mMetaTileEntity.mInventory[i]))<getEUCapacity()){
-                                           increaseStoredEnergyUnits((long)ic2.api.info.Info.itemEnergy.getEnergyValue(mMetaTileEntity.mInventory[i]),false);
-                                           mMetaTileEntity.mInventory[i].stackSize--;
-                                           mInventoryChanged = true;
-                                       }
-                                    }
-                                    if (mMetaTileEntity.mInventory[i].stackSize <= 0) {
-                                        mMetaTileEntity.mInventory[i] = null;
-                                        mInventoryChanged = true;
-                                    }
-                                }
-                            }
-                        }
-                    }
-                case 12:
-                    tCode++;
-                    if (aSideServer) {
-                        if (mMetaTileEntity.rechargerSlotCount() > 0 && getStoredEU() > 0) {
-                            for (int i = mMetaTileEntity.rechargerSlotStartIndex(), k = mMetaTileEntity.rechargerSlotCount() + i; i < k; i++) {
-                                if (getStoredEU() > 0 && mMetaTileEntity.mInventory[i] != null) {
-                                    chargeItem(mMetaTileEntity.mInventory[i]);
-                                    if (mMetaTileEntity.mInventory[i].stackSize <= 0) {
-                                        mMetaTileEntity.mInventory[i] = null;
-                                        mInventoryChanged = true;
-                                    }
-                                }
-                            }
-                        }
-                    }
-                case 13:
-                    tCode++;
-                    updateStatus();
-                    if (!hasValidMetaTileEntity()) {
-                        mRunningThroughTick = false;
-                        return;
-                    }
-                case 14:
-                    tCode++;
-                    mMetaTileEntity.onPostTick(this, mTickTimer);
-                    if (!hasValidMetaTileEntity()) {
-                        mRunningThroughTick = false;
-                        return;
-                    }
-                case 15:
-                    tCode++;
-                    if (aSideServer) {
-                        if (mTickTimer > 20 && cableUpdateDelay == 0) {
-                            generatePowerNodes();
-                            cableUpdateDelay--;
-                        } else {
-                            cableUpdateDelay--;
-                        }
-                        if (mTickTimer % 10 == 0) {
-                            if (mSendClientData) {
-                                NW.sendPacketToAllPlayersInRange(worldObj,
-                                        new GT_Packet_TileEntity(xCoord, (short) yCoord, zCoord, mID,
-                                                mCoverSides[0], mCoverSides[1], mCoverSides[2], mCoverSides[3], mCoverSides[4], mCoverSides[5],
-                                                oTextureData = (byte) ((mFacing & 7) | (mActive ? 8 : 0) | (mRedstone ? 16 : 0) | (mLockUpgrade ? 32 : 0) | (mWorks ? 64 : 0) ),
-                                                oTexturePage = (hasValidMetaTileEntity() && mMetaTileEntity instanceof GT_MetaTileEntity_Hatch) ? ((GT_MetaTileEntity_Hatch) mMetaTileEntity).getTexturePage() : 0,
-                                                oUpdateData = hasValidMetaTileEntity() ? mMetaTileEntity.getUpdateData() : 0,
-                                                oRedstoneData = (byte) (((mSidedRedstone[0] > 0) ? 1 : 0) | ((mSidedRedstone[1] > 0) ? 2 : 0) | ((mSidedRedstone[2] > 0) ? 4 : 0) | ((mSidedRedstone[3] > 0) ? 8 : 0) | ((mSidedRedstone[4] > 0) ? 16 : 0) | ((mSidedRedstone[5] > 0) ? 32 : 0)),
-                                                oColor = mColor),
-                                        xCoord, zCoord);
-                                mSendClientData = false;
-                            }
-                        }
-
-                        if (mTickTimer > 10) {
-                            byte tData = (byte) ((mFacing & 7) | (mActive ? 8 : 0) | (mRedstone ? 16 : 0) | (mLockUpgrade ? 32 : 0)| (mWorks ? 64 : 0));
-                            if (tData != oTextureData)
-                                sendBlockEvent(ClientEvents.CHANGE_COMMON_DATA, oTextureData = tData);
-
-                            tData = mMetaTileEntity.getUpdateData();
-                            if (tData != oUpdateData)
-                                sendBlockEvent(ClientEvents.CHANGE_CUSTOM_DATA, oUpdateData = tData);
-                            if (mMetaTileEntity instanceof GT_MetaTileEntity_Hatch) {
-                                tData = ((GT_MetaTileEntity_Hatch) mMetaTileEntity).getTexturePage();
-                                if (tData != oTexturePage)
-                                    sendBlockEvent(ClientEvents.CHANGE_CUSTOM_DATA, (byte) ((oTexturePage = tData) | 0x80));//set last bit as a flag for page
-                            }
-                            if (mColor != oColor) sendBlockEvent(ClientEvents.CHANGE_COLOR, oColor = mColor);
-                            tData = (byte) (((mSidedRedstone[0] > 0) ? 1 : 0) | ((mSidedRedstone[1] > 0) ? 2 : 0) | ((mSidedRedstone[2] > 0) ? 4 : 0) | ((mSidedRedstone[3] > 0) ? 8 : 0) | ((mSidedRedstone[4] > 0) ? 16 : 0) | ((mSidedRedstone[5] > 0) ? 32 : 0));
-                            if (tData != oRedstoneData)
-                                sendBlockEvent(ClientEvents.CHANGE_REDSTONE_OUTPUT, oRedstoneData = tData);
-                            if (mLightValue != oLightValue) {
-                                worldObj.setLightValue(EnumSkyBlock.Block, xCoord, yCoord, zCoord, mLightValue);
-                                worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord, zCoord);
-                                worldObj.updateLightByType(EnumSkyBlock.Block, xCoord + 1, yCoord, zCoord);
-                                worldObj.updateLightByType(EnumSkyBlock.Block, xCoord - 1, yCoord, zCoord);
-                                worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord + 1, zCoord);
-                                worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord - 1, zCoord);
-                                worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord, zCoord + 1);
-                                worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord, zCoord - 1);
-                                issueTextureUpdate();
-                                sendBlockEvent(ClientEvents.CHANGE_LIGHT, oLightValue = mLightValue);
-                            }
-                        }
-
-                        if (mNeedsBlockUpdate) {
-                            updateNeighbours(mStrongRedstone, oStrongRedstone);
-                            oStrongRedstone = mStrongRedstone;
-                            mNeedsBlockUpdate = false;
-                        }
-                    }
-                default:
-                    tCode = -1;
+                }
             }
-        }
         } catch (Throwable e) {
-            //gregtech.api.util.GT_Log.err.println("Encountered Exception while ticking MetaTileEntity in Step " + (tCode - 1) + ". The Game should've crashed now, but I prevented that. Please report immidietly to GregTech Intergalactical!!!");
             e.printStackTrace(GT_Log.err);
         }
 
@@ -678,6 +632,30 @@ public class BaseMetaTileEntity extends BaseTileEntity implements IGregTechTileE
         }
 
         mWorkUpdate = mInventoryChanged = mRunningThroughTick = false;
+    }
+
+    private void checkDropCover() {
+        for (byte i = 0; i < 6; i++)
+            if (getCoverIDAtSide(i) != 0)
+                if (!mMetaTileEntity.allowCoverOnSide(i, new GT_ItemStack(getCoverIDAtSide(i))))
+                    dropCover(i, i, true);
+    }
+
+    private void sendClientData() {
+        if (mSendClientData) {
+            NW.sendPacketToAllPlayersInRange(
+                worldObj,
+                new GT_Packet_TileEntity(
+                    xCoord, (short) yCoord, zCoord, mID, mCoverSides[0], mCoverSides[1], mCoverSides[2], mCoverSides[3], mCoverSides[4], mCoverSides[5],
+                    oTextureData = (byte) ((mFacing & 7) | (mActive ? 8 : 0) | (mRedstone ? 16 : 0) | (mLockUpgrade ? 32 : 0) | (mWorks ? 64 : 0)),
+                    oTexturePage = (hasValidMetaTileEntity() && mMetaTileEntity instanceof GT_MetaTileEntity_Hatch) ? ((GT_MetaTileEntity_Hatch) mMetaTileEntity).getTexturePage() : 0,
+                    oUpdateData = hasValidMetaTileEntity() ? mMetaTileEntity.getUpdateData() : 0,
+                    oRedstoneData = (byte) (((mSidedRedstone[0] > 0) ? 1 : 0) | ((mSidedRedstone[1] > 0) ? 2 : 0) | ((mSidedRedstone[2] > 0) ? 4 : 0) | ((mSidedRedstone[3] > 0) ? 8 : 0) | ((mSidedRedstone[4] > 0) ? 16 : 0) | ((mSidedRedstone[5] > 0) ? 32 : 0)),
+                    oColor = mColor),
+                xCoord, zCoord
+            );
+            mSendClientData = false;
+        }
     }
 
     @Override
